refactor(tree-view): remove common and use dedicated folder
