+++ lib/attendees/attendees_bloc.dart
@@ -3,7 +3,7 @@ import 'dart:collection';
 
 import 'package:flutter/cupertino.dart';
 import 'package:x_qrcode/api/api_service.dart';
-import 'package:x_qrcode/common/bloc.dart';
+import 'package:x_qrcode/bloc/bloc.dart';
 import 'package:x_qrcode/visitor/model/attendee_model.dart';
 
 enum AttendeesEvents {
