+++ lib/attendees/attendees_screen.dart
@@ -6,8 +6,8 @@ import 'package:flutter/services.dart';
 import 'package:flutter_sticky_header/flutter_sticky_header.dart';
 import 'package:x_qrcode/api/api_service.dart';
 import 'package:x_qrcode/attendees/attendees_bloc.dart';
-import 'package:x_qrcode/attendees/checkin_exception.dart';
-import 'package:x_qrcode/common/circle_gravatar_widget.dart';
+import 'package:x_qrcode/exception/checkin_exception.dart';
+import 'package:x_qrcode/widget/circle_gravatar_widget.dart';
 import 'package:x_qrcode/visitor/model/attendee_model.dart';
 import 'package:x_qrcode/visitor/widget/scan_floating_action_widget.dart';
 import 'package:x_qrcode/visitor/widget/search_input_widget.dart';
