--- lib/attendees/checkin_exception.dart
@@ -1,10 +1,0 @@
-class CheckInException implements Exception {
-  final String message;
-
-  CheckInException(this.message);
-
-  @override
-  String toString() {
-    return message;
-  }
-}
