--- lib/common/bloc.dart
@@ -1,3 +1,0 @@
-abstract class Bloc {
-  void dispose();
-}
