--- lib/common/constants.dart
@@ -1,6 +1,0 @@
-const MODE_CHECK_IN = 'checkIn';
-const MODE_SPONSOR = 'sponsor';
-const MODE = [MODE_CHECK_IN, MODE_SPONSOR];
-
-const ROLE_ADMIN = 'ROLE_ADMIN';
-const ROLE_SPONSOR = 'ROLE_SPONSOR';
