+++ lib/constants.dart
@@ -1,3 +1,10 @@
+const MODE_CHECK_IN = 'checkIn';
+const MODE_SPONSOR = 'sponsor';
+const MODE = [MODE_CHECK_IN, MODE_SPONSOR];
+
+const ROLE_ADMIN = 'ROLE_ADMIN';
+const ROLE_SPONSOR = 'ROLE_SPONSOR';
+
 const APP_NAMESPACE = 'http://x-qrcode.techx.fr';
 
 const STORAGE_KEY_ACCESS_TOKEN = 'st_access_token';
