+++ lib/visitors/visitors_screen.dart
@@ -5,8 +5,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:x_qrcode/api/api_service.dart';
-import 'package:x_qrcode/common/bloc_provider.dart';
-import 'package:x_qrcode/common/circle_gravatar_widget.dart';
+import 'package:x_qrcode/bloc/bloc_provider.dart';
+import 'package:x_qrcode/widget/circle_gravatar_widget.dart';
 import 'package:x_qrcode/visitor/widget/scan_floating_action_widget.dart';
 import 'package:x_qrcode/visitor/visitor_screen.dart';
 import 'package:x_qrcode/visitor/widget/search_input_widget.dart';
